requirements.txt
python-telegram-bot==20.7
beautifulsoup4==4.12.3
requests==2.31.0
aiohttp==3.9.1
lxml==5.1.0
python-dotenv==1.0.0
